refactor(admin): group custom commands module under admin/commands
The custom-command listing/editor page, its state composable and styles lived in separate generic folders. Move Commands.vue, useCommands.ts and commands.scss into one folder so command management can be maintained as a single module.
